+++ client/components/mma/dataPrivacy/DataPrivacyPage.tsx
@@ -1,10 +1,10 @@
 import type { CMP } from '@guardian/consent-management-platform/dist/types';
 import { from } from '@guardian/source-foundations';
 import { useEffect, useState } from 'react';
+import { JsonResponseProcessor } from '@/client/utilities/responseHandlers';
 import { gridItemPlacement } from '../../../styles/grid';
 import { fetchWithDefaultParameters } from '../../../utilities/fetch';
 import {
-	JsonResponseProcessor,
 	LoadingState,
 	useAsyncLoader,
 } from '../../../utilities/hooks/useAsyncLoader';
