+++ client/components/mma/holiday/ExistingHolidayStopActions.tsx
@@ -1,10 +1,10 @@
 import { Button } from '@guardian/source-react-components';
 import { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
+import { JsonResponseProcessor } from '@/client/utilities/responseHandlers';
 import { DATE_FNS_LONG_OUTPUT_FORMAT } from '../../../../shared/dates';
 import { MDA_TEST_USER_HEADER } from '../../../../shared/productResponse';
 import {
-	JsonResponseProcessor,
 	LoadingState,
 	useAsyncLoader,
 } from '../../../utilities/hooks/useAsyncLoader';
